+++ llm_bench/python/llm_run_stable_diffusion_on_linux.sh
@@ -20,12 +20,23 @@ convert_model="python ./llm_bench/python/convert.py --model_id ${original_dir} -
 echo ${convert_model}
 eval ${convert_model}
 echo "convert model ret=$?"
+ret=$?
 wait
+
+if [ ${ret} -ne 0]; then
+    echo "convert model ret=${ret}"
+    exit ${ret}
+end
 
 benchmarking="python ./llm_bench/python/benchmark.py -m ${dst_dir}/pytorch/dldt/FP16/ -pf ./llm_bench/python/prompts/stable-diffusion.jsonl -d cpu -n 1"
 echo ${benchmarking}
 eval ${benchmarking}
-echo "benchmarking ret=$?"
+ret=$?
 
 rm -rf ${original_dir}
 rm -rf ${dst_dir}
+
+if [ ${ret} -ne 0]; then
+    echo "benchmarking ret=${ret}"
+    exit ${ret}
+end
